#7 Setting up the postgres test using a simple table from a new project based in postgres
