+++ SpotWelder.Tests.Common/SpotWelder.Tests.Common.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="NUnit" Version="4.1.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\SpotWelder.Lib\SpotWelder.Lib.csproj" />
+  </ItemGroup>
+
 </Project>
